requirements.txt
-i https://pypi.org/simple
argon2-cffi==23.1.0; python_version >= '3.7'
argon2-cffi-bindings==21.2.0; python_version >= '3.6'
asgiref==3.8.1; python_version >= '3.8'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; python_version >= '3.8'
charset-normalizer==3.4.1; python_version >= '3.7'
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==2.3.0
django==5.2; python_version >= '3.10'
django-axes==7.0.2; python_version >= '3.7'
django-debug-toolbar==5.0.1; python_version >= '3.9'
django-extensions==3.2.3; python_version >= '3.6'
djangorestframework==3.15.2; python_version >= '3.8'
djangorestframework-csv==3.0.2
djangorestframework-simplejwt==5.5.0; python_version >= '3.9'
djangorestframework-xml==2.0.0; python_version >= '3.5'
drf-yasg==1.21.9; python_version >= '3.6'
flake8==7.1.1; python_full_version >= '3.8.1'
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
load-dotenv==0.1.0
mccabe==0.7.0; python_version >= '3.6'
packaging==24.2; python_version >= '3.8'
pipfile==0.0.2
pluggy==1.5.0; python_version >= '3.8'
psycopg[binary]==3.2.4; python_version >= '3.8'
psycopg-binary==3.2.4; python_version >= '3.8'
pycodestyle==2.12.1; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pyjwt==2.9.0; python_version >= '3.8'
pytest==8.3.4; python_version >= '3.8'
pytest-django==4.10.0; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2025.1
pyyaml==6.0.2; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
sqlparse==0.5.3; python_version >= '3.8'
toml==0.10.2; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2'
typing-extensions==4.12.2; python_version >= '3.8'
tzdata==2025.1; python_version >= '2'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.3.0; python_version >= '3.9'
whitenoise==6.9.0; python_version >= '3.9'
